pyproject.toml
[project]
name = "getbooks"
version = "0.1.0"
description = ""
authors = [
    {name = "Luis",email = "luismaiasombra@gmai.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "google-api-python-client (>=2.191.0,<3.0.0)",
    "google-auth (>=2.48.0,<3.0.0)",
    "google-auth-httplib2 (>=0.3.0,<0.4.0)",
    "pymupdf (>=1.27.1,<2.0.0)",
    "pypdf (>=6.7.5,<7.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "supabase (>=2.28.0,<3.0.0)",
    "openai (>=2.24.0,<3.0.0)",
    "google-auth-oauthlib (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
